refactor : LadderLine 패키징 변경, ResultView 메서드 위치 변경
