+++ src/main/java/nextstep/ladder/domain/line/LadderLineGenerator.java
@@ -1,6 +1,5 @@
-package nextstep.ladder.domain.ladderline;
+package nextstep.ladder.domain.line;
 
-import nextstep.ladder.domain.line.Line;
 import nextstep.ladder.domain.point.RightPointStrategy;
 
 import java.util.stream.Collectors;
@@ -8,7 +7,7 @@ import java.util.stream.IntStream;
 
 import static nextstep.ladder.utils.CommonConstant.NUMBER_ZERO;
 
-public class LadderGenerator {
+public class LadderLineGenerator {
 
     public static LadderLine generateLadderLine(int height, int countOfUser, RightPointStrategy rightPointStrategy) {
         return new LadderLine(IntStream.range(NUMBER_ZERO, height)
